feat: ajuste de precisão (double de horas para int segundos)

+++ api/src/main/java/com/nomeacao/api/dto/DadosGrafico.java
@@ -1,6 +1,7 @@
 package com.nomeacao.api.dto;
 
 public record DadosGrafico(
-    String label,  // Ex: "10/12"
-    Double valor   // Ex: 2.5 horas
-) {}
+        String label, // Ex: "10/12"
+        Long valor // Ex: 3600 segundos
+) {
+}
